Split traits and implementations
